% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jecdf.R
\name{densdf}
\alias{densdf}
\title{densdf}
\usage{
densdf(df, nobs)
}
\arguments{
\item{df}{A dataframe of raster cells. Each row should be a single cell.
Each column should be a variable of interest (e.g., elevation, slope, etc).}
}
\value{
A list of dataframes containing the estimated empirical probability
 and cumulative density (mass) functions.
}
\description{
Calculates density dataframes—used internally by other functions.
}
