% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{findOverlap}
\alias{findOverlap}
\title{findOverlap}
\usage{
findOverlap(a, b)
}
\arguments{
\item{a}{A vector of 2 dimensions—e.g., (1,10)}

\item{b}{A vector of 2 dimensions—e.g., (5,9)}
}
\value{
Int. Indicating relationship of boundaries.
}
\description{
Determines the relationship between the bounds of two sets (ranges). Imagine
two sets, 'a' and 'b', where 'a' has bounds represented graphically by []
and 'b' has bounds ()—note this is not intended to represent set notation.
The function returns 1 if 'b' is entirely within 'a', like this: [()]. It
return 2 if [(]); 3 if [](); 4 if ([]); 5 if ([)]; and lastly 6 if ()[]. This
function is used internally if the 'overlap' paramater is True in the 'lamap'
function.
}
