% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parlamap.R
\name{getObservedData}
\alias{getObservedData}
\title{getObservedData}
\usage{
getObservedData(x, rasterdata)
}
\arguments{
\item{x}{Cell index.}

\item{rasterdata}{A raster or stack object from the pacakge 'raster'.}
}
\value{
A single row dataframe with columns corresponding to cell
 coordinates and variables.
}
\description{
Extracts data for a single raster cell from a raster stack.
}
