% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jecdf.R
\name{jdensity}
\alias{jdensity}
\title{jdensity}
\usage{
jdensity(observed, steps, pcdfs, nosupport = NA, partial = T,
  interpolate = F)
}
\arguments{
\item{observed}{Single row dataframe of observed data (will be coerced).
Each column should be one variable that corresponds to one of the cdfs in
pcdfobj.}

\item{pcdfs}{List of cdfs, each corresponding to a relevant predictive
variable. See 'pcdf()'.}

\item{nosupport}{Value to return if the observed value(s) fall outside the
support of the corresponding empirical cdf(s). Defaults to NA.}

\item{partial}{If True, a density of >= 0 will be returned even if there is
only partial overlap between the range (observed+/-steps) and the support
of the corresponding cdf. The default value is True and it is generally
recommended. It means that the function will return the density
(mass) estimate for the portion of a given observed range that does fall
within the support of the corresponding cdf.
Without this option, the endpoint of the range that falls outside the
cdf support will yield the nosupport value, which will mean the entire range
could be disregraded in further calculations.}

\item{interpolate}{If True, the empirical cdf will be interpolated, providing
an estimate of a continuous pdf instead. If False (default), then observed
values that lie between mass density estimates will yield 'nosupport'.}
}
\value{
Vector of densities (masses) corresponding to the variables in
 pcdfobj. The order of the densities (masses) will be the same as the order
 of the corresponding cdfs in pcdfobj.
}
\description{
A wrapper for multicdf() that extracts densities (masses) for a given set
of observed values +/- the vector 'steps'.
}
