% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{knownsitePcdfs}
\alias{knownsitePcdfs}
\title{knownsitePcdfs}
\usage{
knownsitePcdfs(knownsitedf)
}
\arguments{
\item{probs}{A vector of probabilities}

\item{combinations}{A list of matrices containing the the possible
of probabilities. This values in the matrix cells should be the indeces of
the probabilities in the vector 'probs'. Default is Null—so the combinations
will be arranged by the function. To create this combination list, you can
use the 'prepCombinations' convenience function provided in the 'lamap'
package.}
}
\value{
Probability of the union.
}
\description{
Calculates the probability of a union of independent events using the
Law of Total probability and the inclusion-exclusion principle.
}
