% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{numericFactorLevels}
\alias{numericFactorLevels}
\title{numericFactorLevels}
\usage{
numericFactorLevels(x)
}
\arguments{
\item{x}{Table Factor.}
}
\value{
Numeric factor levels.
}
\description{
Returns numeric factor levels from a Table Factor—used internally.
}
