% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jecdf.R
\name{pcdf}
\alias{pcdf}
\title{pcdf}
\usage{
pcdf(traindf, interpolate = F)
}
\arguments{
\item{traindf}{A dataframe of raster cells. Each row should be a single cell.
Each column should be a variable of interest (e.g., elevation, slope, etc).}

\item{inteprolate}{Should the cdf(s) be interpolated to estimate a continuous
pdf? Defaults to False. Ideally, all data should be rounded (effectively
binned) before it's brought into R for lamap analysis. So, interpolation
should not be used. It would undermine one of the fundamental advantages of
of the lamap theory—namely, that multimodal terrain traits could be
important for predictive modelling. Unless the data are very high resolution,
interpolating could erroneously 'smooth out' the very features of the spatial
cdfs that were important to site location.}
}
\value{
List containing pcdfs.
}
\description{
Estimates the empirical probability and cumulative density (mass) functions
for a dataframe of raster cell values. This function is mostly used internally.
}
