% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jecdf.R
\name{pdfIntegral}
\alias{pdfIntegral}
\title{pdfIntegral}
\usage{
pdfIntegral(df)
}
\arguments{
\item{df}{A dataframe with x and y columns. This function is used internally
by the 'pcdf' function.}
}
\value{
An integral estimate (single numeric value).
}
\description{
Calculates a step-wise integral.
}
