% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prepCombinations}
\alias{prepCombinations}
\title{prepCombinations}
\usage{
prepCombinations(n)
}
\arguments{
\item{n}{The number of things to combine in 1–n ways.}
}
\value{
A length n list of matrices.
}
\description{
Prepares a list of combinations (matrices containing the possible
combinations of n).
}
