% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.R
\name{w_exponential}
\alias{w_exponential}
\title{w_exponential}
\usage{
w_exponential(x, param)
}
\arguments{
\item{x}{A vector of distances.}

\item{param}{A vector of weighting function parameters. See primary 'weight'
function.}
}
\value{
A length n list of matrices.
}
\description{
Calculates a vector of weights using an exponential weighting function.
}
