% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.R
\name{w_uniform}
\alias{w_uniform}
\title{w_uniform}
\usage{
w_uniform(x, param)
}
\arguments{
\item{x}{A vector of distances.}

\item{param}{A vector of weighting function parameters. See primary 'weight'
function.}
}
\value{
A length n list of matrices.
}
\description{
Calculates a vector of weights using a uniform weighting function.
}
